Add derive traits and fix clippy warnings
- Add Debug, PartialEq, and Default derives to enums and structs where appropriate
- Replace manual Default implementations with derive attributes for SegmentationType and SegmentationUpidType
- Remove unnecessary type casts in commands.rs and parser.rs
- Fix manual arithmetic check using saturating_sub in parser.rs
- Allow clippy warning for SegmentationDescriptor::new due to necessary parameter count
- Fix needless doctest main wrapper in README.md

+++ src/commands.rs
@@ -211,7 +211,7 @@ pub(crate) fn parse_splice_time(reader: &mut BitReader) -> Result<SpliceTime, io
     let time_specified_flag = reader.read_bslbf(1)? as u8;
     let pts_time = if time_specified_flag == 1 {
         let _reserved = reader.read_bslbf(6)? as u8;
-        Some(reader.read_uimsbf(33)? as u64)
+        Some(reader.read_uimsbf(33)?)
     } else {
         let _reserved = reader.read_bslbf(7)? as u8;
         None
@@ -226,7 +226,7 @@ pub(crate) fn parse_splice_time(reader: &mut BitReader) -> Result<SpliceTime, io
 pub(crate) fn parse_break_duration(reader: &mut BitReader) -> Result<BreakDuration, io::Error> {
     let auto_return = reader.read_bslbf(1)? as u8;
     let reserved = reader.read_bslbf(6)? as u8;
-    let duration = reader.read_uimsbf(33)? as u64;
+    let duration = reader.read_uimsbf(33)?;
     Ok(BreakDuration {
         auto_return,
         reserved,
